08-web-emogo-backend/requirements.txt
fastapi[all]
motor
python-dotenv
uvicorn
